+++ example/android/app/src/main/kotlin/com/mirfut/flutter_tsc_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.mirfut.flutter_tsc_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}